Rework of network API and Server Lua events (#203)
* Refactoring server networking and lua bindings

- Lua print now supports multiple arguments
- finishes adding LuaSystem call functions
- Lua called for connect/disconnect events
- adds TCP no_delay
- ServerMessage header binary serialization
- ServerMessage body length reduced for debugging
- Endpoint tracking
- Cleanup/Fix server protopack loading
- Refactor server logging

* Clean up messaging, start on async client

* Async client and messaging experiment

* Convert to new messaging system

* client more Async (wip)
* message API cleaned up a bit
* added a test suite
* client-server-connection expanded and renamed to run after other tests

* fixes and tests multipart message decoding

* Some cleanup after merge

* Some changes based on feedback and review

+++ client/game.hpp
@@ -56,10 +56,10 @@ private:
 
 	// Frames per second
 	unsigned int frames = 0;
-	float frame_deltas = 0.0;
+	double frame_deltas = 0.0;
 
-	// Elapsed time since last call, same as OS::GetDeltaTime()
-	double GetElapsedTime();
+	// Elapsed time since last call, same as OS::GetDeltaTime(), but as a float
+	float GetElapsedTime();
 	double last_time = 0.0;
 
 	ServerStats stats;
